Fix sign-extension in parse-time disassembly variable computation
Root cause: the DisassemblyPattern::ass_field() method used
i128::from(token_read(bytes)) which zero-extends all token fields.
For signed fields (like RISC-V branch offsets), the sign bit was
lost during the disassembly variable computation at parse time.

This caused all backward branches in RISC-V to go to wrong addresses:
- BEQ x0,x0,-8: target was 0x2FFC instead of 0xFF8
- JAL x1,-4: target was 0x200FFC instead of 0xFFC

Fix: check raw_value_is_signed() in ass_field() and sign-extend
from the exact bit width (8/16/32 via type cast, or arbitrary width
via explicit sign bit check). Same pattern as token_field_as_u64()
and gen_dis_expr_for_lift().

Also fixed ExprValue::DisVar in lower_value(): cast i128 through i64
before u64 to preserve sign for negative displacement values.

RISC-V backward branch tests now hard-assert correct targets.
MIPS backward BEQ remains as known bug (different root cause —
likely MIPS-specific word-offset or delay slot handling).

+++ src/codegen/builder/disassembler/constructor/execution.rs
@@ -1043,9 +1043,11 @@ impl<'a> ExecutionGenerator<'a> {
                 let sz = Self::bytes_from_bits(dv.size.get()) as u32;
                 match self.constructor.dis_fields.get(&dv.id) {
                     Some(name) => {
-                        // Use the local recomputed variable (shadowed from self)
+                        // Use the local recomputed variable (shadowed from self).
+                        // DisVars are i128 — cast through i64 to preserve sign extension
+                        // for negative values (e.g., backward branch displacements).
                         (
-                            quote! { pcode_ir::Varnode::constant(#name as u64, #sz) },
+                            quote! { pcode_ir::Varnode::constant((#name as i64) as u64, #sz) },
                             quote! {},
                         )
                     }
